% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-confint.R
\name{boot.confint}
\alias{boot.confint}
\title{Bootstrap confidence intervals}
\usage{
boot.confint(x.hat, x.star, alpha, boot.method)
}
\arguments{
\item{x.hat}{estimator on the original sample.}

\item{x.star}{vector gathering the estimation on the bootstrapped sample.}

\item{alpha}{1-level of the confidence interval}

\item{boot.method}{bootstrap method.}
}
\value{
A vector of dimension two with the desired confidence interval
}
\description{
\code{boot.confint} computes bootstrap confidence intervals given an estimation on the original sample and on the bootstrap samples
}
