% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-confint.R
\name{confint.LR}
\alias{confint.LR}
\title{Confidence intervals for the Lorenz Regression}
\usage{
\method{confint}{LR}(LR)
}
\arguments{
\item{LR}{Output of a call to \code{\link{Lorenz.Reg}}, where \code{penalty="none"} and \code{Boot.inference=TRUE}.}
}
\value{
A list with two elements
\describe{
   \item{\code{Gi}}{a matrix where each row is a confidence interval for the explained Gini coefficient. The different rows correspond to different bootstrap methods.}
   \item{\code{LR2}}{a matrix where each row is a confidence interval for the Lorenz-\eqn{R^2}. The different rows correspond to different bootstrap methods.}
}
}
\description{
\code{confint.LR} provides confidence intervals for the explained Gini coefficient and Lorenz-R2 for an object of class \code{LR}.
}
\details{
This function is only meant to display the confidence intervals obtained from an already computed LR.
The level and bootstrap methods must be chosen via the arguments \code{alpha} and \code{which.CI} in the \code{Lorenz.Reg} function.
}
\examples{
data(Data.Incomes)
NPLR <- Lorenz.Reg(Income ~ ., data = Data.Incomes, penalty = "none", seed.boot = 123, B = 40, Boot.inference = TRUE)
confint.LR(NPLR)

}
\seealso{
\code{\link{Lorenz.Reg}}
}
