% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.Fitness_cpp}
\alias{.Fitness_cpp}
\title{Computes the fitness used in the GA}
\usage{
.Fitness_cpp(x, Y, X, Z, pi)
}
\arguments{
\item{x}{vector of size (p-1) giving the proposed candidate, where p is the number of covariates}

\item{Y}{vector of size n gathering the response, where n is the sample size}

\item{X}{matrix of dimension (n*p) gathering the covariates}

\item{Z}{vector of size n gathering iid repetitions of a U[0,1]}

\item{pi}{vector of size n gathering the observation weights (notice that sum(pi)=1)}
}
\value{
Fitness of candidate x
}
\description{
Computes the fitness of a candidate in the genetic algorithm displayed in function Lorenz.GA.cpp
}
