% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.frac_rank_cpp}
\alias{.frac_rank_cpp}
\title{Computes fractional ranks}
\usage{
.frac_rank_cpp(x, pi)
}
\arguments{
\item{x}{vector of size n gathering the values for which the fractional rank should be computed}

\item{pi}{vector of size n gathering the observation weights (notice that sum(pi)=1)}
}
\value{
Fractional rank related to vector x
}
\description{
Computes the vector of fractional ranks related to a given vector
}
\keyword{internal}
