% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqExplained.R
\name{ineqExplained.lm}
\alias{ineqExplained.lm}
\title{Explained inequality metrics for (generalized) linear models}
\usage{
\method{ineqExplained}{lm}(object, type = c("Gini.explained", "Lorenz-R2"), ...)
}
\arguments{
\item{object}{An object of S3 class \code{"lm"}.}

\item{type}{Character string specifying the type of inequality metric to retrieve. Options are \code{"Gini.explained"} (for the explained Gini coefficient) and \code{"Lorenz-R2"} (for the Lorenz-\eqn{R^2}).}

\item{...}{Additional arguments passed to \code{\link{Gini.coef}}.}
}
\value{
A numeric value representing the requested inequality metric.
}
\description{
Retrieves the explained Gini coefficient or the Lorenz-\eqn{R^2} from an object of class \code{"lm"}.
}
