% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqExplained.R
\name{ineqExplained}
\alias{ineqExplained}
\title{Retrieve a measure of explained inequality from a model}
\usage{
ineqExplained(object, type = c("Gini.explained", "Lorenz-R2"), ...)
}
\arguments{
\item{object}{An object for which the inequality metrics should be extracted.}

\item{type}{Character string specifying the type of inequality metric to retrieve. Options are \code{"Gini.explained"} for the explained Gini coefficient or \code{"Lorenz-R2"} for the Lorenz-\eqn{R^2}.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
The requested inequality metric.
}
\description{
This generic function extracts a measure of explained inequality, such as the explained Gini coefficient or the Lorenz-R2, from a fitted model object.
}
\examples{
## For examples see example(Lorenz.Reg)

}
\seealso{
\code{\link{Lorenz.Reg}}
}
