% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lorenz-curve.R
\name{Lorenz.curve}
\alias{Lorenz.curve}
\title{Concentration curve of \emph{y} with respect to \emph{x}}
\usage{
Lorenz.curve(y, x = y, graph = F)
}
\arguments{
\item{y}{variable of interest.}

\item{x}{variable to use for the ranking. By default \eqn{x=y}, and the obtained concentration curve is the Lorenz curve of \emph{y}.}

\item{graph}{whether a graph of the obtained concentration curve should be traced. Default value is FALSE.}
}
\value{
A function corresponding to the estimated Lorenz or concentration curve. If \code{graph} is TRUE, the curve is also plotted.
}
\description{
\code{Lorenz.curve} computes the concentration curve index of a vector \emph{y} with respect to another vector \emph{x}.
If \emph{y} and \emph{x} are identical, the obtained concentration curve boils down to the Lorenz curve.
}
\examples{
data(Data.Incomes)
# We first compute the Lorenz curve of Income
Y <- Data.Incomes$Income
Lorenz.curve(y = Y, graph = TRUE)
# Then we compute the concentration curve of Income with respect to Age
X <- Data.Incomes$Age
Lorenz.curve(y = Y, x = X, graph = TRUE)

}
\seealso{
\code{\link{Lorenz.graphs}}, \code{\link{Gini.coef}}
}
