% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA-estimation.R
\name{Lorenz.GA.cpp}
\alias{Lorenz.GA.cpp}
\title{Estimates the parameter vector in Lorenz regression using a genetic algorithm}
\usage{
Lorenz.GA.cpp(YX_mat, popSize = 50, maxiter = 1500, run = 150, parallel = F)
}
\arguments{
\item{YX_mat}{A matrix with the first column corresponding to the response vector, the remaining ones being the explanatory variables.}

\item{popSize}{Size of the population of candidates in the genetic algorithm. Default value is 50.}

\item{maxiter}{Maximum number ot iterations in the genetic algorithm. Default value is 1500.}

\item{run}{Number of iterations without improvement in the best fitness necessary for the algorithm to stop. Default value is 150.}

\item{parallel}{Whether parallel computing should be used in the genetic algorithm. Default value is FALSE.}
}
\value{
A list with several components:
\describe{
   \item{\code{sol}}{the estimated vector of parameters.}
   \item{\code{LR2}}{the Lorenz-\eqn{R^2} of the regression.}
   \item{\code{Gi.expl}}{the estimated explained Gini coefficient.}
   \item{\code{niter}}{number of iterations attained by the genetic algorithm.}
   \item{\code{fit}}{value attained by the fitness function at the optimum.}
}
}
\description{
\code{Lorenz.GA.cpp} estimates the vector of parameters in Lorenz regression using the unit-norm normalization
It also returns the Lorenz-\eqn{R^2} of the regression.
}
\details{
The genetic algorithm is solved using function \code{\link[GA]{ga}} from the \emph{GA} package. The fitness function is coded in Rcpp to speed up computation time.
When discrete covariates are introduced and ties occur in the index, they are randomly broken, as advised in Section 3 of Heuchenne and Jacquemain (2020)
}
\section{References}{

Heuchenne, C. and A. Jacquemain. 2020. “Inference for monotone single-index conditional means: a Lorenz regression approach”
}

\examples{
data(Data.Incomes)
YX_mat <- cbind(Data.Incomes$Income, Data.Incomes$Age, Data.Incomes$Work.Hours)
Lorenz.GA.cpp(YX_mat, popSize = 40)

}
\seealso{
\code{\link{Lorenz.Reg}}, \code{\link[GA]{ga}}
}
