% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA-population.R
\name{Lorenz.Population}
\alias{Lorenz.Population}
\title{Defines the population used in the genetic algorithm}
\usage{
Lorenz.Population(object)
}
\arguments{
\item{object}{An object of class "\code{ga}", resulting from a call to function \code{ga}.}
}
\value{
A matrix of dimension \code{object@popSize} times the number of explanatory variables minus one, gathering the initial population.
}
\description{
\code{Lorenz.Population} creates the initial population of the genetic algorithm used to solve the Lorenz regression.
}
\details{
Note that this population produces an initial solution ensuring a unit norm.
}
\seealso{
\code{\link{Lorenz.GA}}
}
\keyword{internal}
