% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-matrix-PLR.R
\name{model_matrix_PLR}
\alias{model_matrix_PLR}
\title{Design matrix in the Penalized Lorenz Regression}
\usage{
model_matrix_PLR(mt, mf)
}
\arguments{
\item{mt}{Model terms}

\item{mf}{Model frame}
}
\value{
The design matrix
}
\description{
\code{model_matrix_PLR} is a utilitary function that provides the design matrix for the Penalized Lorenz Regression
}
\details{
This function ensures that the design matrix is constructed according to the requirements of the PLR.
In PLR, one must exclude the intercept and use one-hot encoding for all variables, except when binary
}
\keyword{internal}
