% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLR-OOB.R
\name{PLR.OOB}
\alias{PLR.OOB}
\title{Computes the OOB for the Penalized Lorenz Regression}
\usage{
PLR.OOB(y, x, weights, theta.boot)
}
\arguments{
\item{y}{the vector of responses}

\item{x}{the design matrix (after data management steps, i.e. standardization and transformations of the categorical covariates into binaries)}

\item{weights}{vector of sample weights related to the validation sample. By default, each observation is given the same weight.}

\item{theta.boot}{list of matrices. Each element of the list correspond to a value of the tuning parameter. The columns of the matrices correspond to values of the penalty parameters. The rows correspond to the different covariates.}
}
\value{
A list of vectors gathering the OOB-score. Each element of the list corresponds to a value of the tuning parameter and each element of the vector corresponds to a value of the penalization parameter.
}
\description{
\code{PLR.OOB} computes the out-of-bag score obtained for a specific validation sample and associated to a list of parameters obtained by the Penalized Lorenz Regression.
}
