% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLR-traceplot.R
\name{PLR.traceplot}
\alias{PLR.traceplot}
\title{Trace plot of a Penalized Lorenz Regression}
\usage{
PLR.traceplot(PLR, type.var = NULL)
}
\arguments{
\item{PLR}{Output of a call to \code{\link{Lorenz.Reg}}, where \code{penalty!="none"}.}
}
\value{
A plot where the horizontal axis is -log(lambda), lambda being the value of the regularization parameter, and where the vertical axis gives the size of the coefficient attached to each covariate.
}
\description{
\code{PLR.traceplot} displays the trace plot of a Penalized Lorenz Regression, i.e. it displays the evolution of the estimated coefficient attached to each covariate as a function of the regularization parameter.
}
\examples{
data(Data.Incomes)
PLR <- Lorenz.Reg(Income ~ ., data = Data.Incomes, penalty = "SCAD", lambda.choice = "BIC", eps = 0.005, nfolds = 5)
PLR.traceplot(PLR)

}
\seealso{
\code{\link{Lorenz.Reg}}
}
