% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-print.R
\name{print.LR}
\alias{print.LR}
\title{Printing method for the Lorenz Regression}
\usage{
\method{print}{LR}(x, ...)
}
\arguments{
\item{x}{Output of a call to \code{\link{Lorenz.Reg}}, where \code{penalty="none"}.}

\item{...}{Additional arguments.}
}
\value{
No return value, called for printing an object of class \code{LR} to the console
}
\description{
\code{print.LR} prints the arguments and estimated coefficients of an object of class \code{LR}.
}
\examples{
data(Data.Incomes)
NPLR <- Lorenz.Reg(Income ~ ., data = Data.Incomes, penalty = "none")
print(NPLR)

}
\seealso{
\code{\link{Lorenz.Reg}}
}
