% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLR-print-summary.R
\name{print.summary.PLR}
\alias{print.summary.PLR}
\title{Printing method for the summary of a Penalized Lorenz Regression}
\usage{
\method{print}{summary.PLR}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{...}{Additional arguments}

\item{object}{An object of class \code{"summary.PLR"}.}
}
\value{
No return value, called for printing an object of class \code{"PLR"} to the console.
}
\description{
\code{print.summary.PLR} provides a printing method for an object of class \code{"summary.PLR"}.
}
\examples{
data(Data.Incomes)
NPLR <- Lorenz.Reg(Income ~ ., data = Data.Incomes, penalty = "none")
summary(NPLR)

}
\seealso{
\code{\link{summary.PLR}}
}
