% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-predict.R
\name{residuals.LR}
\alias{residuals.LR}
\alias{residuals.LR_boot}
\title{Residuals for the Lorenz regression}
\usage{
\method{residuals}{LR}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"LR"}.}

\item{...}{Additional arguments passed to the function \code{\link{Rearrangement.estimation}}.}
}
\value{
A vector of residuals.
}
\description{
\code{residuals} provides residuals for an object of class \code{"LR"}.
}
\details{
Computing residuals entail to estimate the link function of the single-index model. This is done via the function \code{\link{Rearrangement.estimation}}.
}
\examples{
## For examples see example(Lorenz.Reg) and example(Lorenz.boot)

}
\seealso{
\code{\link{Lorenz.Reg}}, \code{\link{Rearrangement.estimation}}
}
