% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runif_seed.R
\name{runif_seed}
\alias{runif_seed}
\title{Generates a sample of uniform random variables with a specific seed}
\usage{
runif_seed(n, min = 0, max = 1, seed = NULL)
}
\arguments{
\item{n}{the sample size}

\item{seed}{the seed to use}
}
\value{
A vector with the generated random variables
}
\description{
\code{runif_seed} generates a vector of uniform(0,1) random variables with a specific seed. The seed is only used locally.
}
\keyword{internal}
