% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-plot.R
\name{autoplot.LR}
\alias{autoplot.LR}
\alias{plot.LR}
\title{Plots for the Lorenz Regression}
\usage{
\method{autoplot}{LR}(x, ...)

\method{plot}{LR}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"LR"}.}

\item{...}{Additional arguments passed to \code{\link{Lorenz.graphs}}.}
}
\value{
\code{autoplot.LR} returns a \code{ggplot} object representing the Lorenz curve of the response and the concentration curve of the response with respect to the estimated index. \code{plot.LR} directly displays this plot.
}
\description{
\code{autoplot.LR} generates a plot for an object of class \code{"LR"} and returns it as a \code{ggplot} object.
The \code{plot.LR} method is a wrapper around \code{autoplot.LR} that directly displays the plot,
providing a more familiar interface for users accustomed to base R plotting.
}
\examples{
## For examples see example(Lorenz.Reg)

}
\seealso{
\code{\link{Lorenz.Reg}}
}
