% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-plot.R
\name{autoplot.LR}
\alias{autoplot.LR}
\alias{plot.LR}
\alias{autoplot.LR_boot}
\alias{plot.LR_boot}
\title{Plots for the Lorenz regression}
\usage{
\method{autoplot}{LR}(object, ...)

\method{plot}{LR}(x, ...)
}
\arguments{
\item{object}{An object of class \code{"LR"}.}

\item{...}{Additional arguments passed to \code{\link{Lorenz.graphs}}.}

\item{x}{An object of class \code{"LR"}.}
}
\value{
\code{autoplot} returns a \code{ggplot} object representing the Lorenz curve of the response and the concentration curve of the response with respect to the estimated index. \code{plot} directly displays this plot.
}
\description{
\code{autoplot} generates a plot for an object of class \code{"LR"} and returns it as a \code{ggplot} object.
The \code{plot} method is a wrapper around \code{autoplot} that directly displays the plot,
providing a more familiar interface for users accustomed to base R plotting.
}
\examples{
## For examples see example(Lorenz.Reg)

}
\seealso{
\code{\link{Lorenz.Reg}}
}
