% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-coef.R
\name{coef.LR}
\alias{coef.LR}
\alias{coef.LR_boot}
\title{Estimated coefficients for the Lorenz regression}
\usage{
\method{coef}{LR}(object, ...)
}
\arguments{
\item{object}{An object of S3 class \code{"LR"}.}

\item{...}{Additional arguments.}
}
\value{
a vector gathering the estimated coefficients
}
\description{
Provides the estimated coefficients for an object of class \code{"LR"}.
}
\examples{
## For examples see example(Lorenz.Reg)

}
\seealso{
\code{\link{Lorenz.Reg}}
}
