% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-confint.R
\name{confint.LR}
\alias{confint.LR}
\title{Confidence intervals for the Lorenz Regression}
\usage{
\method{confint}{LR}(
  object,
  parm = c("Gini", "LR2", "theta"),
  level = 0.95,
  type = c("norm", "basic", "perc"),
  bias.corr = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{"LR_boot"}.}

\item{parm}{A logical value determining whether the confidence interval is computed for the explained Gini coefficient, for the Lorenz-R2 or for the vector of theta coefficients. Possible values are \code{"Gini"} (default, for the explained Gini),\code{"LR2"} (for the Lorenz-R2) and \code{"theta"} (for the vector theta).}

\item{level}{A numeric giving the level of the confidence interval. Default value is 0.95.}

\item{type}{A character string specifying the bootstrap method. Possible values are \code{"norm"}, \code{"basic"} and \code{"perc"}. For more information, see the argument \code{type} of the function \code{\link{boot.ci}} from the \code{\link{boot}} library.}

\item{bias.corr}{A logical determining whether bias correction should be performed. Only used if \code{type="norm"}. Default is \code{TRUE}.}

\item{...}{Additional arguments.}
}
\value{
The desired confidence interval.
If \code{parm="Gini"} or \code{parm="LR2"}, the output is a vector.
If \code{parm="theta"}, it is a matrix where each row corresponds to a different coefficient.
}
\description{
\code{confint.LR} provides bootstrap confidence intervals for the explained Gini coefficient, Lorenz-R2 and theta vector for an object of class \code{"LR_boot"}.
}
\examples{
\donttest{
# The following piece of code might take several minutes
data(Data.Incomes)
set.seed(123)
Data <- Data.Incomes[sample(1:nrow(Data.Incomes),50),]
NPLR <- Lorenz.Reg(Income ~ ., data = Data, penalty = "none",
                   seed.boot = 123, B = 40, Boot.inference = TRUE)
confint(NPLR)
}

}
\seealso{
\code{\link{Lorenz.boot}}, \code{\link[boot]{boot.ci}}
}
