% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLR-confint.R
\name{confint.PLR}
\alias{confint.PLR}
\title{Confidence intervals for the Penalized Lorenz Regression}
\usage{
\method{confint}{PLR}(PLR)
}
\arguments{
\item{PLR}{Output of a call to \code{\link{Lorenz.Reg}}, where \code{penalty!="none"} and \code{Boot.inference=TRUE}.}
}
\value{
A list with two elements
\describe{
   \item{\code{Gi}}{a list of confidence intervals for the explained Gini coefficient. Each element of the list correspond to a bootstrap method.}
   \item{\code{LR2}}{a list of confidence intervals for the Lorenz-\eqn{R^2}. Each element of the list correspond to a bootstrap method.}
}
}
\description{
\code{confint.PLR} provides confidence intervals for the explained Gini coefficient and Lorenz-R2 for an object of class \code{PLR}.
}
\details{
This function is only meant to display the confidence intervals obtained from an already computed PLR.
The level and bootstrap methods must be chosen via the arguments \code{alpha} and \code{which.CI} in the \code{Lorenz.Reg} function.
}
\examples{
data(Data.Incomes)
PLR <- Lorenz.Reg(Income ~ ., data = Data.Incomes, h.grid = nrow(Data.Incomes)^(-1/5.5), penalty = "SCAD", eps = 0.01, seed.boot = 123, B = 50, Boot.inference = TRUE)
confint.PLR(PLR)

}
\seealso{
\code{\link{Lorenz.Reg}}
}
