% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLR-confint.R
\name{confint.PLR_boot}
\alias{confint.PLR_boot}
\title{Confidence intervals for the Penalized Lorenz Regression}
\usage{
\method{confint}{PLR_boot}(
  object,
  parm = c("Gini", "LR2"),
  level = 0.95,
  type = c("norm", "basic", "perc"),
  which.pars = NULL,
  bias.corr = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{"PLR_boot"}.}

\item{parm}{A character string determining whether the confidence interval is computed for the explained Gini coefficient or for the Lorenz-R2. Possible values are \code{"Gini"} (default, for the explained Gini) and \code{"LR2"} (for the Lorenz-R2)}

\item{level}{A numeric giving the level of the confidence interval. Default value is 0.95.}

\item{type}{A character string specifying the bootstrap method. Possible values are \code{"norm"}, \code{"basic"} and \code{"perc"}. For more information, see the argument \code{type} of the function \code{\link{boot.ci}} from the \emph{boot} library.}

\item{which.pars}{A vector of size 2 specifying the index of the tuning parameter (first element) and the index of the penalty parameter (second element) that should be selected.
Default is \code{NULL}, in which case the parameters are selected by the available methods : BIC (always), bootstrap (if \code{object} inherits from the \code{PLR_boot} class) and cross-validation (if \code{object} inherits from the \code{PLR_cv} class).}

\item{bias.corr}{A logical determining whether bias correction should be performed. Only used if \code{type="norm"}. Default is \code{TRUE}.}
}
\value{
The desired confidence interval.
If \code{which.pars=NULL}, the output is a matrix where each row corresponds to a selection method.
Otherwise, the tuning and penalty parameters are specified by the vector \code{which.pars} and the output is a vector.
}
\description{
\code{confint.PLR_boot} provides bootstrap confidence intervals for the explained Gini coefficient and Lorenz-R2 for an object of class \code{"PLR_boot"}.
}
\examples{
## For examples see example(Lorenz.boot)

}
\seealso{
\code{\link{Lorenz.boot}}, \code{\link[boot]{boot.ci}}
}
