% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-ineq_explained.R
\name{ineqExplained}
\alias{ineqExplained}
\alias{ineqExplained.LR}
\title{Retrieve a measure of explained inequality from a model}
\usage{
ineqExplained(object, type = c("Gini.explained", "Lorenz-R2"), ...)

\method{ineqExplained}{LR}(object, type = c("Gini.explained", "Lorenz-R2"), ...)
}
\arguments{
\item{object}{An object of S3 class \code{"LR"}.}

\item{type}{Character string specifying the type of inequality metric to retrieve. Options are \code{"Gini.explained"} and \code{"Lorenz-R2"}.}

\item{...}{Additional arguments.}
}
\value{
The requested inequality metric.

A numeric value representing the requested inequality metric.
}
\description{
This generic function extracts a measure of explained inequality, such as the explained Gini coefficient or the Lorenz-R2, from a fitted model object.
}
\section{Methods (by class)}{
\itemize{
\item \code{ineqExplained(LR)}: Extract inequality metrics from an \code{"LR"} object

\code{ineqExplained.LR} retrieves the explained Gini coefficient or the Lorenz-\eqn{R^2} from an object of class \code{"LR"}.

}}
\examples{
## For examples see example(Lorenz.Reg)

}
\seealso{
\code{\link{Lorenz.Reg}}
}
