% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lorenz-bootstrap.R
\name{Lorenz.boot}
\alias{Lorenz.boot}
\title{Bootstrap tests or confidence intervals for Lorenz regression}
\usage{
Lorenz.boot(
  formula,
  data,
  Lorenz.est = NULL,
  standardize = T,
  testorCI = "both",
  which.CI = c("Basic", "Perc", "Param"),
  which.vars.test,
  alpha = 0.05,
  B,
  parallel = F
)
}
\arguments{
\item{formula}{A formula object of the form \emph{response} ~ \emph{other_variables}.}

\item{data}{A data frame containing the variables displayed in the formula.}

\item{Lorenz.est}{Output of a Lorenz.Reg on the original dataset. Default value is NULL, in which case the vector of parameters is estimated internally.}

\item{standardize}{should the variables be standardized before the estimation process? Default value is TRUE}

\item{testorCI}{whether we should do a bootstrap test of joint significance or a CI or both. Possibles values are "test", "CI" or "both".}

\item{which.CI}{vector gathering the methods to construct the CI, among Basic Bootstrap ("Basic"), Percentile Bootstrap ("Perc") and Parametric Bootstrap
("Param"). Default value is c("Basic","Perc","Param"), meaning that the three methods are used.}

\item{which.vars.test}{names of the variables to consider for the test of joint significance.}

\item{alpha}{level of the CI and/or tests.}

\item{B}{the number of bootstrap iterations.}

\item{parallel}{whether parallel computing should be used. Default value is FALSE.}
}
\value{
A list containing at most the following elements
\describe{
    \item{\code{Basic.theta}}{matrix with p rows and 2 columns (lower and upper bounds) providing the basic bootstrap CI for theta.}
    \item{\code{Param.theta}}{idem for the parametric bootstrap.}
    \item{\code{Perc.theta}}{idem for the percentile bootstrap}
    \item{\code{Basic.Gi}}{vector providing the lower and upper bounds of the basic bootstrap CI for the explained Gini}
    \item{\code{Param.Gi}}{idem for the parametric bootstrap.}
    \item{\code{Perc.Gi}}{idem for the percentile bootstrap.}
    \item{\code{Param.pvalues}}{p-values indicating the significance of each coefficient, computed using parametric bootstrap.}
    \item{\code{pvalue}}{p-value of the required test of joint significance}
    \item{\code{U}}{observed value of the test statistic}
}
}
\description{
\code{Lorenz.boot} produces inference on the vector of parameters of the Lorenz regression.
More specifically, this function can be used to obtain individual confidence intervals for each element of the parameter vector.
It can also be used to perform a test of joint significance of several covariates.
We refer the user to Heuchenne and Jacquemain (2020) for a rigorous exposition of the procedures.
}
\section{References}{

Heuchenne, C. and A. Jacquemain. 2020. “Inference for monotone single-index conditional means: a Lorenz regression approach”
}

\examples{
# # Example A) takes approx 7.8 minutes to complete.
# # Example B) takes approx 3.9 minutes to complete.
# # In practice, we advise to set B at least 200
# # Example A)
# a <- Sys.time()
# data(Data.Incomes)
# Lorenz.boot(Income ~ ., data = Data.Incomes, B = 50, which.vars.test = c("Age","Seniority"))
# b <- Sys.time() - a
# b
# # Computation time can be sped up using parallel computing
# c <- Sys.time()
# Lorenz.boot(Income ~ ., data = Data.Incomes, B = 50,
#             which.vars.test = c("Age","Seniority"), parallel = T)
# d <- Sys.time() - c
# d

}
\seealso{
\code{\link{Lorenz.Reg}}
}
