% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lorenz-regression.R
\name{Lorenz.Reg}
\alias{Lorenz.Reg}
\title{Undertakes a Lorenz regression}
\usage{
Lorenz.Reg(formula, data, standardize = T, weights = NULL, ...)
}
\arguments{
\item{formula}{A formula object of the form \emph{response} ~ \emph{other_variables}.}

\item{data}{A data frame containing the variables displayed in the formula.}

\item{standardize}{Should the variables be standardized before the estimation process? Default value is TRUE.}

\item{weights}{vector of sample weights. By default, each observation is given the same weight.}

\item{...}{Additional parameters corresponding to arguments passed in \code{\link{Lorenz.GA.cpp}}}
}
\value{
A list with several components:
\describe{
   \item{\code{theta}}{the estimated vector of parameters (on the original scale, even if \code{standardize} is TRUE).}
   \item{\code{expl.Gini}}{the estimated explained Gini coefficient}
   \item{\code{LorenzR2}}{the Lorenz-\eqn{R^2} of the regression.}
   \item{\code{MRS}}{the matrix of estimated marginal rates of substitution. More precisely, if we want the MRS of X1 (numerator) with respect to X2 (denominator),
   we should look for row corresponding to X1 and column corresponding to X2}.
   \item{\code{Fit}}{A data frame containing the response (first column) and the estimated index (second column).}
}
}
\description{
\code{Lorenz.Reg} performs the Lorenz regression of a response with respect to several covariates.
}
\examples{
data(Data.Incomes)
Lorenz.Reg(Income ~ Age + Work.Hours, data = Data.Incomes)


}
\seealso{
\code{\link{Lorenz.GA.cpp}}, \code{\link{Lorenz.boot}}
}
