% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lorenz-regression.R
\name{Lorenz.Reg}
\alias{Lorenz.Reg}
\title{Fits a Lorenz regression}
\usage{
Lorenz.Reg(
  formula,
  data,
  weights,
  na.action,
  penalty = c("none", "SCAD", "LASSO"),
  grid.arg = c("h", "SCAD.nfwd", "eps", "kernel", "a", "gamma"),
  grid.value = NULL,
  lambda.list = NULL,
  ...
)
}
\arguments{
\item{formula}{An object of class "\code{\link{formula}}" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{An optional data frame, list or environment (or object coercible by \code{\link{as.data.frame}} to a data frame) containing the variables in the model. If not found in \code{data}, the variables are taken from \code{environment(formula)}, typically the environment from which \code{Lorenz.Reg} is called.}

\item{weights}{An optional vector of sample weights to be used in the fitting process. Should be \code{NULL} or a numeric vector.}

\item{na.action}{A function which indicates what should happen when the data contain \code{NA}s. The default is set by the \code{na.action} setting of \code{\link{options}}, and is \code{\link{na.fail}} if that is unset. The 'factory-fresh' default is \code{\link{na.omit}}. Another possible value is \code{NULL}, no action. Value \code{\link{na.exclude}} can be useful.}

\item{penalty}{A character string specifying the type of penalty on the size of the estimated coefficients of the single-index model.
The default value is \code{"none"}, in which case a non-penalized Lorenz regression is fitted using \code{\link{Lorenz.GA}}.
Other possible values are \code{"LASSO"} and \code{"SCAD"}, in which case a penalized Lorenz regression is fitted using \code{\link{Lorenz.FABS}} or \code{\link{Lorenz.SCADFABS}} respectively.}

\item{grid.arg}{A character string specifying the tuning parameter for which a grid is to be constructed, see Details.}

\item{grid.value}{A numeric vector specifying the grid values, see Details.}

\item{lambda.list}{Technical argument used inside the function \code{\link{Lorenz.boot}}.}

\item{...}{Additional parameters corresponding to arguments passed in \code{\link{Lorenz.GA}}, \code{\link{Lorenz.FABS}} or \code{\link{Lorenz.SCADFABS}}, depending on the argument chosen in \code{penalty}.}
}
\value{
An object of class \code{"LR"} for the non-penalized Lorenz regression or of class \code{"PLR"} for a penalized Lorenz regression.

For both classes, several methods are available. The function \code{summary} is used to summarize the model fits.
Information on the coefficient of the single-index model is obtained via \code{coef}.
The method \code{predict} is used to predict either the response or the index of the model.
A visual representation of explained inequality through Lorenz curves is provided with the method \code{plot}.

The object of class \code{"LR"} is a list containing the following components:
\describe{
   \item{\code{theta}}{The estimated vector of parameters.}
   \item{\code{Gi.expl}}{The estimated explained Gini coefficient.}
   \item{\code{LR2}}{The Lorenz-\eqn{R^2} of the regression.}
   \item{\code{MRS}}{The matrix of estimated marginal rates of substitution. More precisely, if we want the MRS of \eqn{X_1} (numerator) with respect to \eqn{X_2} (denominator),
   we should look for row corresponding to \eqn{X_1} and column corresponding to \eqn{X_2}.}
   \item{\code{index}}{A vector gathering the estimated index.}
}
For the Penalized Lorenz Regression, the grid parameter (i.e. the value on the grid \code{grid.value}) and the penalization parameter (lambda) are chosen optimally via the BIC method.
The object of class \code{"PLR"} is a list containing the same components as previously, and in addition :
\describe{
   \item{\code{path}}{A list where the different elements correspond to the values of the grid parameter. Each element is a matrix where the first line displays the vector of lambda values. The second and third lines display the evolution of the Lorenz-\eqn{R^2} and explained Gini coefficient along that vector. The next lines display the evolution of the BIC score. The remaining lines display the evolution of the estimated coefficients of the single-index model.}
   \item{\code{lambda.idx}}{the index of the optimal lambda obtained by the BIC method}
   \item{\code{grid.idx}}{the index of the optimal grid parameter obtained by the BIC method.}
}
In both cases, the list also provides technical information, such as the specified \code{formula}, \code{weights} and \code{call}, as well as the design matrix \code{x} and the response vector \code{y}.
}
\description{
\code{Lorenz.Reg} fits the Lorenz regression of a response with respect to several covariates.
}
\details{
In the penalized case, the model is fitted for a grid of values of two parameters : the penalty parameter (lambda) and one tuning parameter specified by the arguments \code{grid.arg} and \code{grid.value}.
The possibles values for \code{grid.arg} are tuning parameters of the functions \code{\link{Lorenz.FABS}} and \code{\link{Lorenz.SCADFABS}} : \code{''h''} (the default), \code{''SCAD.nfwd''},\code{''eps''}, \code{''kernel''}, \code{''a''} and \code{''gamma''}.
The values for the grid are specified with \code{grid.value}. The default is \code{NULL}, in which case no grid is constructed
}
\section{References}{

Heuchenne, C. and A. Jacquemain (2022). Inference for monotone single-index conditional means: A Lorenz regression approach. \emph{Computational Statistics & Data Analysis 167(C)}.

Jacquemain, A., C. Heuchenne, and E. Pircalabelu (2024). A penalised bootstrap estimation procedure for the explained Gini coefficient. \emph{Electronic Journal of Statistics 18(1) 247-300}.
}

\examples{
data(Data.Incomes)
set.seed(123)
data <- Data.Incomes[sample(1:200,50),]
# 1. Non-penalized regression
NPLR <- Lorenz.Reg(Income ~ ., data = Data.Incomes, penalty = "none", popSize = 30)
# 2. Penalized regression
PLR <- Lorenz.Reg(Income ~ ., data = Data.Incomes, penalty = "SCAD",
                  eps = 0.02, grid.arg = "h",
                  grid.value=c(0.5,1,2)*nrow(Data.Incomes)^(-1/5.5))
# Print method
print(NPLR)
print(PLR)
# Summary method
summary(NPLR)
summary(PLR)
# Coef method
coef(NPLR)
coef(PLR)
# Predict method
## One can predict either the index or the response
predict(NPLR,type="response")
predict(PLR,type="response")
# Plot method
plot(NPLR)
plot(PLR)
## Traceplot of the penalized coefficients
plot(PLR,type="traceplot")

}
\seealso{
\code{\link{Lorenz.GA}}, \code{\link{Lorenz.SCADFABS}}, \code{\link{Lorenz.FABS}}, code{\link{Lorenz.boot}}
}
