% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-plot.R
\name{plot.LR}
\alias{plot.LR}
\title{Plots for the Lorenz Regression}
\usage{
\method{plot}{LR}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"LR"}.}

\item{...}{Additional arguments}
}
\value{
The Lorenz curve of the response and concentration curve of the response with respect to the estimated index. The graph is as an object of class \code{"ggplot"}.
}
\description{
\code{plot.LR} provides plots for an object of class \code{"LR"}.
}
\examples{
data(Data.Incomes)
NPLR <- Lorenz.Reg(Income ~ ., data = Data.Incomes, penalty = "none")
plot(NPLR)

}
\seealso{
\code{\link{Lorenz.Reg}}
}
