% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-plot.R
\name{plot.LR}
\alias{plot.LR}
\title{Plots for the Unpenalized Lorenz Regression}
\usage{
\method{plot}{LR}(x, ...)
}
\arguments{
\item{x}{Output of a call to \code{\link{Lorenz.Reg}}, where \code{penalty=="none"}.}

\item{...}{Additional arguments}
}
\value{
The Lorenz curve of the response and concentration curve of the response with respect to the estimated index
}
\description{
\code{plot.LR} provides plots for an object of class \code{LR}.
}
\examples{
data(Data.Incomes)
NPLR <- Lorenz.Reg(Income ~ ., data = Data.Incomes, penalty = "none")
plot(NPLR)

}
\seealso{
\code{\link{Lorenz.Reg}}
}
