% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLR-CV.R
\name{PLR.CV}
\alias{PLR.CV}
\title{Determines the regularization parameter (lambda) in a PLR via cross-validation}
\usage{
PLR.CV(
  formula,
  data,
  penalty = "SCAD",
  PLR.est = NULL,
  standardize = T,
  weights = NULL,
  eps,
  nfolds = 10,
  seed.CV = NULL,
  parallel = F,
  ...
)
}
\arguments{
\item{formula}{A formula object of the form \emph{response} ~ \emph{other_variables}.}

\item{data}{A data frame containing the variables displayed in the formula.}

\item{penalty}{penalty used in the Penalized Lorenz Regression. Possible values are "SCAD" (default) or "LASSO".}

\item{PLR.est}{Output of a call to \code{\link{PLR.wrap}} corresponding to the estimation of the Penalized Lorenz Regression on the full sample. Default value is NULL in which case the estimation on the full sample is run internally.}

\item{standardize}{Should the variables be standardized before the estimation process? Default value is TRUE.}

\item{weights}{vector of sample weights. By default, each observation is given the same weight.}

\item{eps}{Step size in the FABS or SCADFABS algorithm. Default value is 0.005.}

\item{nfolds}{Number of folds. Default value is 10.}

\item{seed.CV}{Should a specific seed be used in the definition of the folds. Default value is NULL in which case no seed is imposed.}

\item{parallel}{Whether parallel computing should be used to distribute the \code{nfolds} computations on different CPUs. Default value is FALSE.}

\item{...}{Additional parameters corresponding to arguments passed in \code{\link{Lorenz.SCADFABS}} or \code{\link{Lorenz.FABS}} depending on the argument chosen in penalty.}
}
\value{
A list with two components
\describe{
   \item{\code{val}}{vector indicating the CV-score for each value of lambda.}
   \item{\code{best}}{index where the optimum is attained.}
}
}
\description{
\code{PLR.CV} undertakes k-fold cross-validation for a Penalized Lorenz Regression. It returns the CV-score associated to each value of the regularization parameter and the index of the optimum.
}
\section{References}{

Heuchenne, C. and A. Jacquemain (2022). Inference for monotone single-index conditional means: A Lorenz regression approach. \emph{Computational Statistics & Data Analysis 167(C)}.
}

\examples{
YX_mat <- Data.Incomes
PLR <- PLR.wrap(YX_mat, eps=0.01)
PLR.CV(Income ~ ., Data.Incomes, PLR.est = PLR, eps = 0.01, nfolds = 5)

}
\seealso{
\code{\link{Lorenz.Reg}}, \code{\link{PLR.wrap}}, \code{\link{Lorenz.FABS}}, \code{\link{Lorenz.SCADFABS}}
}
