% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLR-normalize.R
\name{PLR.normalize}
\alias{PLR.normalize}
\title{Re-normalizes the estimated coefficients of a penalized Lorenz regression}
\usage{
PLR.normalize(object)
}
\arguments{
\item{object}{An object of S3 class \code{"PLR"}. The object must also contain a \code{"theta"} slot with the estimated coefficients}
}
\value{
A matrix of re-normalized coefficients.
}
\description{
\code{PLR.normalize} transforms the estimated coefficients of a penalized Lorenz regression to match the model where the first category of each categorical variable is omitted.
}
\seealso{
\code{\link{Lorenz.Reg}}
}
\keyword{internal}
