% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-predict.R
\name{predict.LR}
\alias{predict.LR}
\title{Prediction for the Lorenz regression}
\usage{
\method{predict}{LR}(object, newdata, type = c("index", "response"), ...)
}
\arguments{
\item{object}{An object of class \code{"LR"}.}

\item{type}{A character string indicating the type of prediction. Possible values are \code{"response"} and \code{"index"} (the default).
In the first case, the prediction estimates the conditional expectation of the response given the covariates.
In the second case, the prediction estimates only the index of the single-index model.}

\item{...}{Additional arguments passed to the function \code{\link{Rearrangement.estimation}}.}
}
\value{
a vector gathering the predictions
}
\description{
\code{predict.LR} provides predictions for an object of class \code{LR}.
}
\details{
If \code{type="Response"}, the link function of the single-index model must be estimated. This is done via the function \code{\link{Rearrangement.estimation}}.
}
\examples{
data(Data.Incomes)
NPLR <- Lorenz.Reg(Income ~ ., data = Data.Incomes, penalty = "none")
predict(NPLR)

}
\seealso{
\code{\link{Lorenz.Reg}}, \code{\link{Rearrangement.estimation}}
}
