% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLR-predict.R
\name{predict.PLR}
\alias{predict.PLR}
\title{Prediction for the Penalized Lorenz regression}
\usage{
\method{predict}{PLR}(object, newdata, type = c("index", "response"), pars.idx = NULL, ...)
}
\arguments{
\item{object}{An object of S3 class \code{"PLR"}.}

\item{newdata}{An optional data frame in which to look for variables with which to predict. If omitted, the original data are used.}

\item{type}{A character string indicating the type of prediction. Possible values are \code{"response"} and \code{"index"} (the default).
In the first case, the prediction estimates the conditional expectation of the response given the covariates.
In the second case, the prediction estimates only the index of the single-index model.}

\item{pars.idx}{A vector of size 2 specifying the index of the grid parameter (first element) and the index of the penalty parameter (second element) that should be selected.
Default is \code{NULL}, in which case the parameters are selected by the available methods : BIC (always), bootstrap (if \code{object} inherits from the \code{PLR_boot} class) and cross-validation (if \code{object} inherits from the \code{PLR_cv} class).}

\item{...}{Additional arguments passed to the function \code{\link{Rearrangement.estimation}}.}
}
\value{
a vector gathering the predictions.
If the object has also class \code{"PLR_boot"} and/or \code{"PLR_cv"}, the output is a matrix, where each column corresponds to a selection method.
}
\description{
\code{predict.PLR} provides predictions for an object of class \code{"PLR"}.
}
\details{
If \code{type="response"}, the link function of the single-index model must be estimated. This is done via the function \code{\link{Rearrangement.estimation}}.
}
\examples{
## For examples see example(Lorenz.Reg), example(Lorenz.boot) and example(PLR.CV)

}
\seealso{
\code{\link{Lorenz.Reg}}, \code{\link{Rearrangement.estimation}}
}
