% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLR-predict.R
\name{predict.PLR}
\alias{predict.PLR}
\alias{predict.PLR_boot}
\alias{predict.PLR_cv}
\title{Prediction for the Penalized Lorenz regression}
\usage{
\method{predict}{PLR}(object, newdata, type = c("index", "response"), pars.idx = "BIC", ...)
}
\arguments{
\item{object}{An object of S3 class \code{"PLR"}. The object might also have S3 classes \code{"PLR_boot"} and/or \code{"PLR_cv"} (both inherit from class \code{"PLR"})}

\item{newdata}{An optional data frame in which to look for variables with which to predict. If omitted, the original data are used.}

\item{type}{A character string indicating the type of prediction. Possible values are \code{"response"} and \code{"index"} (the default).
In the first case, the prediction estimates the conditional expectation of the response given the covariates.
In the second case, the prediction estimates only the index of the single-index model.}

\item{pars.idx}{What grid and penalty parameters should be used for parameter selection. Either a character string specifying the selection method, where the possible values are:
\itemize{
   \item \code{"BIC"} (default) - Always available.
   \item \code{"Boot"} - Available if \code{object} inherits from \code{"PLR_boot"}.
   \item \code{"CV"} - Available if \code{object} inherits from \code{"PLR_cv"}.
}
Or a numeric vector of length 2, where the first element is the index of the grid parameter and the second is the index of the penalty parameter.}

\item{...}{Additional arguments passed to the function \code{\link{Rearrangement.estimation}}.}
}
\value{
a vector gathering the predictions.
}
\description{
Provides predictions for an object of class \code{"PLR"}.
}
\details{
If \code{type="response"}, the link function of the single-index model must be estimated. This is done via the function \code{\link{Rearrangement.estimation}}.
}
\examples{
## For examples see example(Lorenz.Reg), example(Lorenz.boot) and example(PLR.CV)

}
\seealso{
\code{\link{Lorenz.Reg}}, \code{\link{Rearrangement.estimation}}
}
