% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-print.R
\name{print.LR}
\alias{print.LR}
\alias{print.LR_boot}
\title{Printing method for the Lorenz regression}
\usage{
\method{print}{LR}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{"LR"}.}

\item{digits}{The number of significant digits to be passed.}

\item{...}{Additional arguments.}
}
\value{
No return value, called for printing an object of class \code{"LR"} to the console.
}
\description{
Prints the arguments, explained Gini coefficient and estimated coefficients of an object of class \code{"LR"}.
}
\examples{
## For examples see example(Lorenz.Reg)

}
\seealso{
\code{\link{Lorenz.Reg}}
}
