% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLR-print.R
\name{print.PLR}
\alias{print.PLR}
\title{Printing method for the Penalized Lorenz Regression}
\usage{
\method{print}{PLR}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{"PLR"}.}

\item{digits}{The number of significant digits to be passed.}

\item{...}{Additional arguments.}
}
\value{
No return value, called for printing an object of class \code{"PLR"} to the console.
}
\description{
\code{print.PLR} prints the arguments, explained Gini coefficient and estimated coefficients of an object of class \code{"PLR"}.
}
\examples{
## For examples see example(Lorenz.Reg), example(Lorenz.boot) and example(PLR.CV)

}
\seealso{
\code{\link{Lorenz.Reg}}
}
