% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-print-summary.R
\name{print.summary.LR}
\alias{print.summary.LR}
\title{Printing method for the summary of a Lorenz Regression}
\usage{
\method{print}{summary.LR}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{...}{Additional arguments}

\item{object}{An object of class \code{"summary.LR"}.}
}
\value{
No return value, called for printing an object of class \code{"LR"} to the console.
}
\description{
\code{print.summary.LR} provides a printing method for an object of class \code{"summary.LR"}.
}
\examples{
data(Data.Incomes)
NPLR <- Lorenz.Reg(Income ~ ., data = Data.Incomes, penalty = "none")
summary(NPLR)

}
\seealso{
\code{\link{summary.LR}}
}
