% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-print-summary.R
\name{print.summary.LR}
\alias{print.summary.LR}
\title{Printing method for the summary of a Lorenz Regression}
\usage{
\method{print}{summary.LR}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{"summary.LR"}.}

\item{digits}{Number of significant digits to be passed.}

\item{signif.stars}{Logical determining whether p-values should be also encoded visually. See the help of the function \code{\link{printCoefmat}} for more information.}

\item{...}{Additional arguments passed to the function \code{\link{print}}.}
}
\value{
No return value, called for printing an object of class \code{"LR"} to the console.
}
\description{
\code{print.summary.LR} provides a printing method for an object of class \code{"summary.LR"}.
}
\examples{
## For examples see example(Lorenz.Reg) and example(Lorenz.boot)

}
\seealso{
\code{\link{summary.LR}}
}
