% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLR-print-summary.R
\name{print.summary.PLR}
\alias{print.summary.PLR}
\title{Printing method for the summary of a Penalized Lorenz Regression}
\usage{
\method{print}{summary.PLR}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{"summary.PLR"}.}

\item{digits}{Number of significant digits to be passed.}

\item{...}{Additional arguments passed to the function \code{\link{print}}.}
}
\value{
No return value, called for printing an object of class \code{"summary.PLR"} to the console.
}
\description{
\code{print.summary.PLR} provides a printing method for an object of class \code{"summary.PLR"}.
}
\examples{
## For examples see example(Lorenz.Reg), example(Lorenz.boot) and example(PLR.CV)

}
\seealso{
\code{\link{summary.PLR}}
}
