% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLR-summary.R
\name{summary.PLR}
\alias{summary.PLR}
\title{Summary for the Penalized Lorenz Regression}
\usage{
\method{summary}{PLR}(object, renormalize = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{"PLR"}.}

\item{...}{Additional arguments}
}
\value{
An object of class \code{"summary.PLR"}, containing the following elements:
\describe{
   \item{\code{call}}{The matched call.}
   \item{\code{Gi.expl}}{The estimated explained Gini coefficient.}
   \item{\code{LR2}}{The Lorenz-\eqn{R^2} of the regression.}
   \item{\code{coefficients}}{A matrix providing information on the estimated coefficients. The first column gives the estimates.
   If the class of \code{object} contains \code{LR_boot}, bootstrap inference was performed and the matrix contains further information. The second column is the boostrap standard error. The third column is the z-value. Finally, the last column is the p-value.}
}
}
\description{
\code{summary.PLR} provides a summary for an object of class \code{"PLR"}.
}
\details{
The inference provided in the \code{coefficients} matrix is obtained by using the asymptotic normality and estimating the asymptotic variance via bootstrap.
}
\examples{
data(Data.Incomes)
NPLR <- Lorenz.Reg(Income ~ ., data = Data.Incomes, penalty = "none")
summary(NPLR)

}
\seealso{
\code{\link{Lorenz.Reg}}, \code{\link{Lorenz.boot}}
}
