% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLR-summary.R
\name{summary.PLR}
\alias{summary.PLR}
\title{Summary for the Penalized Lorenz Regression}
\usage{
\method{summary}{PLR}(object, renormalize = TRUE, pars.idx = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{"PLR"}.}

\item{renormalize}{A logical value determining whether the coefficient vector should be re-normalized to match the representation where the first category of each categorical variable is omitted. Default value is TRUE}

\item{pars.idx}{A vector of size 2 specifying the index of the grid parameter (first element) and the index of the penalty parameter (second element) that should be selected.
Default is \code{NULL}, in which case the parameters are selected by the available methods : BIC (always), bootstrap (if \code{object} inherits from the \code{PLR_boot} class) and cross-validation (if \code{object} inherits from the \code{PLR_cv} class).}

\item{...}{Additional arguments}
}
\value{
An object of class \code{"summary.PLR"}, containing the following elements:
\describe{
   \item{\code{call}}{The matched call.}
   \item{\code{ineq}}{The table of explained inequality. The first column displays the explained Gini coefficient, the second displays the Gini coefficient of the response and the last column displays the Lorenz-R2. Each row corresponds to a different selection method.}
   \item{\code{coefficients}}{A matrix providing information on the estimated coefficients. Each column corresponds to a different selection method}
}
}
\description{
\code{summary.PLR} provides a summary for an object of class \code{"PLR"}.
}
\examples{
## For examples see example(Lorenz.Reg), example(Lorenz.boot) and example(PLR.CV)

}
\seealso{
\code{\link{Lorenz.Reg}}, \code{\link{Lorenz.boot}}
}
