% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLR-summary.R
\name{summary.PLR}
\alias{summary.PLR}
\title{Summary for the Penalized Lorenz Regression}
\usage{
\method{summary}{PLR}(PLR, renormalize = TRUE)
}
\arguments{
\item{PLR}{Output of a call to \code{\link{Lorenz.Reg}}, where \code{penalty!="none"}.}

\item{renormalize}{whether the coefficient vector should be re-normalized to match the representation where the first category of each categorical variable is omitted. Default value is TRUE}
}
\value{
A summary displaying two tables: a summary of the model and the estimated coefficients.
}
\description{
\code{summary.PLR} provides a summary for an object of class \code{PLR}.
}
\examples{
data(Data.Incomes)
PLR <- Lorenz.Reg(Income ~ ., data = Data.Incomes, penalty = "SCAD", lambda.choice = c("BIC","CV"), eps = 0.005, seed.CV = 123, nfolds = 5)
summary.PLR(PLR)

}
\seealso{
\code{\link{Lorenz.Reg}}
}
